+++ includes/class-wpglobus-wpseo.php
@@ -391,7 +391,7 @@ class WPGlobus_WPSEO {
 			$wpseotitle = get_post_meta( $post->ID, '_yoast_wpseo_title', true );
 			$focuskw    = get_post_meta( $post->ID, '_yoast_wpseo_focuskw', true ); 			
 			foreach ( WPGlobus::Config()->open_languages as $language ) {
-				$url = WPGlobus_Utils::localize_url( $permalink['url'], $language ); 	?>
+				$url = apply_filters( 'wpglobus_wpseo_permalink', WPGlobus_Utils::localize_url( $permalink['url'], $language ), $language ); ?>
 				<div id="wpseo-tab-<?php echo $language; ?>" class="wpglobus-wpseo-general"
 				     data-language="<?php echo $language; ?>"
 				     data-url-<?php echo $language; ?>="<?php echo $url; ?>"