+++ includes/js/wpglobus-admin.js
@@ -174,7 +174,8 @@ var WPGlobusDialogApp;
 			var option = {
 				id: null,
 				dialogTitle: '',
-				style: ''
+				style: '',
+				styleTextareaWrapper: '',
 			}
 			if ( 'string' == typeof(elem) ) {
 				option.id = elem;	
@@ -236,7 +237,7 @@ var WPGlobusDialogApp;
 			$(clone).insertAfter('#'+id);
 			if ( 'TEXTAREA' == node.nodeName ) {
 				$('#wpglobus-'+id).addClass( 'wpglobus-textarea-'+id );
-				$('.wpglobus-textarea-'+id).wrapAll( '<div class="wpglobus-textarea-wrapper"></div>' );
+				$('.wpglobus-textarea-'+id).wrapAll( '<div class="wpglobus-textarea-wrapper" style="'+option.styleTextareaWrapper+'"></div>' );
 			}
 			$(document).on('change', '#wpglobus-'+id, function(){
 				var $t = $(this), 
@@ -966,8 +967,61 @@ jQuery(document).ready(function () {
                 }
 
 				// wp_editor word count
-				if ( typeof(wpWordCount) != 'undefined' ) {
+				if ( typeof wp.utils != 'undefined' && typeof wp.utils.WordCounter != 'undefined' ) {							
+					// from WordPress 4.3 @see c:\cygwin\home\www.wpg.dev\wp-admin\js\post.js
+					$.each(WPGlobusCoreData.enabled_languages, function(i,l){
+						( function( $, counter, l ) {
+							$( function() {
+								if ( l == WPGlobusCoreData.default_language ) {
+									var $content = $( '#content' ),
+										$count = $( '#wp-word-count' ).find( '.word-count' );							
+								} else {
+									var $content = $( '#content_'+l ),
+										$count = $( '#wp-word-count-'+l ).find( '.word-count-'+l );
+								}	
+									
+								var	prevCount = 0,
+									contentEditor;
 
+								function update() {
+									var text, count;
+
+									if ( ! contentEditor || contentEditor.isHidden() ) {
+										text = $content.val();
+									} else {
+										text = contentEditor.getContent( { format: 'raw' } );
+									}
+
+									count = counter.count( text );
+
+									if ( count !== prevCount ) {
+										$count.text( count );
+									}
+
+									prevCount = count;
+								}
+
+								$( document ).on( 'tinymce-editor-init', function( event, editor ) {
+									if ( editor.id !== 'content' ) {
+										return;
+									}
+
+									contentEditor = editor;
+
+									editor.on( 'nodechange keyup', _.debounce( update, 1000 ) );
+								} );
+
+								$content.on( 'input keyup', _.debounce( update, 1000 ) );
+
+								update();
+							} );
+						} )( jQuery, new wp.utils.WordCounter(), l );
+					});
+					
+				}
+				
+				if ( typeof(wpWordCount) != 'undefined' ) {
+					// wordpress 4.2.4 and earlier
 					var last = 0,
 						ls = WPGlobusCoreData.open_languages,
 						$d = this.$document,